started on guest and user page routes
