+++ src/user.tsx
@@ -1,0 +1,3 @@
+export const UserPage = () => {
+  return <div>This is the user page.</div>;
+};
